+++ app/controllers/note.js
